+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix staking address display [#204]
+
 ## [0.19.0] - 2023-09-20
 
 ### Added
@@ -426,6 +430,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implementation of `Store` trait from `wallet-core`
 - Implementation of `State` and `Prover` traits from `wallet-core`
 
+[#204]: https://github.com/dusk-network/wallet-cli/issues/204
 [#182]: https://github.com/dusk-network/wallet-cli/issues/182
 [#198]: https://github.com/dusk-network/wallet-cli/issues/198
 [#176]: https://github.com/dusk-network/wallet-cli/issues/176